backend/requirements.txt
fastapi>=0.109.2
uvicorn>=0.27.1
sqlalchemy>=2.0.25
asyncpg>=0.29.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
pydantic>=2.6.1
pydantic-settings>=2.1.0
python-dotenv>=1.0.1
python-multipart>=0.0.9
redis>=5.0.1
locust>=2.24.0
uvloop>=0.19.0
httptools>=0.6.1
torch
xgboost
shap
numpy
scikit-learn
